pyproject.toml
[project]
name = "book-loop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==3.2.2",
    "customtkinter>=5.2.2",
    "email-validator>=2.3.0",
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "keyring>=25.7.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
